LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_touch-action-svg-none-test_touch-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Events touch-action attribute support

Test Description: Try to scroll element DOWN then RIGHT starting your touch inside of the circle. Tap Complete button under the rectangle when done.

Note: this test is for touch only



touch-action: none

The following pointer types were detected: (none).


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN touch-action attribute test in SVG
